+++ cloud/org.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 
 	secretsapi "github.com/earthly/cloud-api/secrets"
-	"github.com/golang/protobuf/jsonpb"
 	"github.com/pkg/errors"
 )
 
@@ -25,15 +24,6 @@ type OrgPermissions struct {
 	User  string
 	Path  string
 	Write bool
-}
-
-// OrgInvitation can be used to invite a user to become a member in an org.
-type OrgInvitation struct {
-	Name       string
-	Email      string
-	Permission string
-	Message    string
-	OrgName    string
 }
 
 // ListOrgs lists all orgs a user has permission to view.
@@ -190,36 +180,6 @@ func (c *client) GetOrgID(ctx context.Context, orgName string) (string, error) {
 		}
 	}
 	return "", errors.Errorf("org not found: %s", orgName)
-}
-
-// InviteToOrg sends an email invitation to a user and asks for them to join an org.
-func (c *client) InviteToOrg(ctx context.Context, invite *OrgInvitation) (string, error) {
-	u := "/api/v1/invitations"
-
-	req := &secretsapi.CreateInvitationRequest{
-		Name:       invite.Name,
-		OrgName:    invite.OrgName,
-		Email:      invite.Email,
-		Permission: invite.Permission,
-		Message:    invite.Message,
-	}
-
-	status, body, err := c.doCall(ctx, http.MethodPost, u, withAuth(), withJSONBody(req))
-	if err != nil {
-		return "", err
-	}
-
-	if status != http.StatusOK {
-		return "", errors.Errorf("failed to send email invite: %s", body)
-	}
-
-	res := &secretsapi.CreateInvitationResponse{}
-	err = jsonpb.UnmarshalString(body, res)
-	if err != nil {
-		return "", err
-	}
-
-	return res.Token, nil
 }
 
 func getOrgFromPath(path string) (string, bool) {
